# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/couriers": {
            "get": {
                "description": "Retrieves a list of all available couriers.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Couriers"
                ],
                "summary": "Get all couriers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/API.CourierResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a new courier by providing the courier's name.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Couriers"
                ],
                "summary": "Create a new courier",
                "parameters": [
                    {
                        "description": "Courier name",
                        "name": "courier",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/API.CreateCourierRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/API.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/couriers/{id}": {
            "get": {
                "description": "Retrieves a specific courier by providing the courier ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Couriers"
                ],
                "summary": "Get courier by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Courier ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/API.CourierResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates an existing courier by providing courier ID, name, and status.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Couriers"
                ],
                "summary": "Update courier by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Courier ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated courier data",
                        "name": "courier",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/API.UpdateCourierRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/API.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a specific courier by providing the courier ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Couriers"
                ],
                "summary": "Delete courier by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Courier ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/API.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/API.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "API.CourierResponse": {
            "type": "object",
            "properties": {
                "courier_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "API.CreateCourierRequest": {
            "type": "object",
            "required": [
                "courier_name"
            ],
            "properties": {
                "courier_name": {
                    "type": "string"
                }
            }
        },
...
        "API.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "API.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "API.UpdateCourierRequest": {
            "type": "object",
            "required": [
                "courier_name",
                "status"
            ],
            "properties": {
                "courier_name": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
...
    }
}